Preparation of the project main pages for development.
