+++ finder/lib/pages/favorite_screen.dart
@@ -6,8 +6,39 @@ class FavoriteScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      color: Colors.yellowAccent,
+    return SafeArea(
+        child: Scaffold(
+          appBar: AppBar(
+            title: Text(
+              "Gespeichert", //Username Nutzername
+              style: Theme.of(context).textTheme.titleLarge!.copyWith(),
+            ),
+            actions: [
+              Padding(
+                padding: const EdgeInsets.all(8.0),
+                child: Container(
+                  width: 36,
+                  height: 36,
+                  decoration: BoxDecoration(
+                    borderRadius: BorderRadius.circular(36),
+                    color: Colors.black,
+                  ),
+                  child: Center(
+                    child: Text(
+                      "F",  //Initialien
+                      style: Theme.of(context).textTheme.bodyMedium!.copyWith(
+                        color: Theme.of(context).colorScheme.primary,
+                        fontSize: 27,
+                        fontWeight: FontWeight.bold,
+                      ),
+                    ),
+                  ),
+                ),
+              ),
+            ],
+          ),
+
+        ),
     );
   }
 }
